[HBASE-10568] Adding MemstoreBloomFilter which is an in memory bloom filter
Author: manukranthk

Summary: This diff adds MemstoreBloomFilter which is an in memory bloom filter against the Memstore. This is introduced to add RowKeyPrefixBloomFilter in Memstore and store the bloom filter for Rowkey Prefixes. The underlying bloom filter used is ByteBloomFilter. Since this bloom filter needs to be accessed from different threads, the class needs to be thread safe. This diff also incorporates the changes related to making the ByteBloomFilter class thread safe.

Test Plan: Unit test. Currently I have a basic unit test which tests the common case. Will add more unit tests in the next revision.

Reviewers: adela, liyintang, aaiyer

Reviewed By: liyintang

CC: hbase-eng@, gauravm, arice

Differential Revision: https://phabricator.fb.com/D1152434

Task ID: 3624202

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1569877 13f79535-47bb-0310-9956-ffa450edef68