[HBASE-10578] [0.89-fb] Choose the correct KV from multiple scanners
Author: aaiyer

Summary:
When multiple scanners have the same KV, HBase should pick the "newest" one.
However, if one of the "current" scanner has the same KV, it seems to get
preference over the other scanners in the heap -- THIS IS WRONG.

This issue has been causing Relja problems.  The diff adds a unit test to make
sure that bulk loads correctly. And fixes the issue.

Test Plan: Added a unit test

Reviewers: gauravm, liyintang, rshroff

Reviewed By: rshroff

CC: hbase-dev@, rpetrovic, daviddeng

Differential Revision: https://phabricator.fb.com/D1183013

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1570674 13f79535-47bb-0310-9956-ffa450edef68