[HBASE-10647] Fix the bug of not decrease from size in updateColumnValue when removing the old columns.
Author: daviddeng

Summary:
In `MemStore.updateColumnValue`, when removing old values, `size` is not decreased, but local variable `addedSize` is decreased, so size is greater than real value. This leads to the problem of negative value for `HRegion.memstoreSize`.

Add a new case in `TestMemStore` to test it.

Test Plan: Run `TestMemStore`

Reviewers: manukranthk, aaiyer, rshroff, liyintang

Reviewed By: aaiyer

CC: hbase-eng@, arice

Differential Revision: https://phabricator.fb.com/D1197954

Task ID: 3831205

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1573210 13f79535-47bb-0310-9956-ffa450edef68