[HBASE-9930] Bug fix in HConnectionManager isMeta function.
Author: manukranthk

Summary: Catching an exception and finally removing the element from the isMeta stack.

Test Plan: Tested on tsh053. We should probably replace this stack using a boolean being passed throughout the interface/or a better design to identify the fact that we are looking for the meta table and that it needs to be prioritized.

Reviewers: adela, gauravm

Reviewed By: gauravm

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D1232327

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1584169 13f79535-47bb-0310-9956-ffa450edef68