[HBASE-10905][0.89-fb] Be able to query a region location through the (Thrift)HRegionInterface
Author: gauravm

Summary:
In order to write a simplistic non-Java client (PHP in our case), where we don't want to push the region locating code to the client side, we can let the Region Server do it for us.

The client can pick up an Region Server in the regionservers tier, and get the location for any row in the table. The client is also able to request the server to disregard any previous cached location if it needs to.

Test Plan: TestGetRegionLocation and other unit tests

Reviewers: manukranthk, fan, rshroff, myang, daviddeng

Reviewed By: manukranthk

CC: hbase-eng@, andrewcox

Differential Revision: https://phabricator.fb.com/D1257376

Task ID: 4058577

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1585806 13f79535-47bb-0310-9956-ffa450edef68