[Master] Fix TestRegionChecker
Author: liyintang

Summary: There was a bug in the TestRegionChecker. In the unit tests, the test function tries to verify whether avDetDayAfter contains all the re-assigned regions. However, the ROOT region will not be included in such collection based on the design of the HRegionChecker. Therefore, the fix is to exclude the ROOT region from the checking.

Test Plan: will run this unit tests multiple times to verify

Reviewers: rshroff

Reviewed By: rshroff

CC: hbase-dev@

Differential Revision: https://phabricator.fb.com/D1276146

Task ID: 3297398

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1588115 13f79535-47bb-0310-9956-ffa450edef68