[HBASE-11021] HBaseFsck Fix for Table Hole Reporting
Author: achayes

Summary:
HBaseFsck would not report when multiple regions went missing. Instead, it
would report the first missing region it found in the meta table and exit.
The new implementation now lists all ranges of [startkey, endkey] that are
missing from the meta table, which should allow for easier debugging.

Test Plan:
TestHBaseFsck is a short test file written to test this functionality. It
sets up a minicluster, creates a table, and then deletes a bunch of rows
from the table and runs the checker. Since the checker does not throw errors
on failure, the output logs have to be checked manually to see if there was
any error. The result looks like this:

https://www.facebook.com/pxlcld/l7xF

Reviewers: manukranthk

Reviewed by: manukranthk

CC:

Differential Revision: https://phabricator.fb.com/D1266615

Task ID: #3675218

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1589229 13f79535-47bb-0310-9956-ffa450edef68