[HBASE-9704] Do not retry on Hadoop port on exceptions while creating a thrift client
Author: manukranthk

Summary: We added this functionality such that we will be able to connect to regionservers while in a rolling restart state. Since this assumption does not hold anymore. We should remove this retry and simply throw in case of a failure.

Test Plan: Not a major change. Removing a retry which doesn't work currently

Reviewers: aaiyer, gauravm, adela

Reviewed By: gauravm

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D1290661

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1589821 13f79535-47bb-0310-9956-ffa450edef68