[HBASE-9704] Create a copy of the conf in ClientLocalScanner since we are modifying it
Author: manukranthk

Summary: In ClientLocalScanner.openStoresOnClient, we modify the conf to set the fs.default.name to be the value obtained from the hbase cluster the htable is create for. But, doing so will modify the conf passed in. In case the user intends to use the config later, it might have unknown effects.

Test Plan: Trivial change. Hot fix for search005

Reviewers: daviddeng

Reviewed By: daviddeng

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D1325499

Task ID: 4317015

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1594433 13f79535-47bb-0310-9956-ffa450edef68