[HBASE-10343] Fix TestHRegionSeqidTransition
Author: rshroff

Summary:
The test would fail whenever the 2/3 region servers which are stopped
has the ROOT region server. The reason it fails it because of timeout in
relocating the ROOT region server. The locateMetaInRoot is suppose to
try to locate the META region, but in case the ROOT is moved, it takes
some time, until PFFE kicks in for that server, where is rediscovers the
root location. The flow is correct, it's just that the PFFE threshold is
60 secs, which might cause the test case to timeout in some cases.

The fix is reducing the threshold, so that PFFE kicks in faster.

Test Plan: run it 25 times. It should pass

Reviewers: manukranthk, everyoung, liyintang, daviddeng

Reviewed By: manukranthk

CC: hbase-dev@

Differential Revision: https://phabricator.fb.com/D1316334

Task ID: 4237582

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1594660 13f79535-47bb-0310-9956-ffa450edef68