[HBASE-8500] Fix Rolling Restart to handle scenarios where RegionServer and Master are using different versions of HRegionInfo
Author: rshroff

Summary:
During the table schema change, we update the META region and then
restart the regions one by one in a throttled manner. Incase we fail during
this time, the Master and the RegionServer have different HRegionInfo versions
of the same object.

This causes confusion and failures on the Master side to not reassign the
Region back to the favored node and also on the RollingRestart side, where it fails
to lookup the favored nodes for this region.

The fix is on the RollingRestart side, to do a serial scan in case the get() fails.

Test Plan: test on t043, rolling restart o/p https://phabricator.fb.com/P10344177

Reviewers: aaiyer, liyintang, manukranthk

Reviewed By: manukranthk

CC: hbase-dev@, manukranthk, daviddeng

Differential Revision: https://phabricator.fb.com/D1327052

Task ID: 4311266

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1594661 13f79535-47bb-0310-9956-ffa450edef68